feat: Show all the lessons with the correct format
Now the user is able to iterate through the lessons by clicking the Next button.
The lessons are shown with taking into consideration the availability of the interaction or not.
Done layout is displayed now at the end of the lessons.
Testing code is added to test the added logic by showing mock lessons.